+++ forge/src/main/java/com/christofmeg/justenoughbreeding/emi/TransformationCategoryEMI.java
@@ -1,10 +1,11 @@
 package com.christofmeg.justenoughbreeding.emi;
 
 import com.christofmeg.justenoughbreeding.CommonConstants;
-import com.christofmeg.justenoughbreeding.recipe.TrustingRecipe;
+import com.christofmeg.justenoughbreeding.recipe.TransformationRecipe;
 import dev.emi.emi.api.recipe.EmiRecipe;
 import dev.emi.emi.api.recipe.EmiRecipeCategory;
 import dev.emi.emi.api.recipe.EmiRecipeSorting;
+import dev.emi.emi.api.render.EmiTexture;
 import dev.emi.emi.api.stack.EmiIngredient;
 import dev.emi.emi.api.stack.EmiStack;
 import dev.emi.emi.api.widget.WidgetHolder;
@@ -16,16 +17,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 @SuppressWarnings("removal")
-public class TrustingCategoryEMI extends AbstractRecipeCategoryEMI {
+public class TransformationCategoryEMI extends AbstractRecipeCategoryEMI {
 
-    private final TrustingRecipe recipe;
-    public static EmiRecipeCategory TYPE = new EmiRecipeCategory(
-        new ResourceLocation(CommonConstants.MOD_ID + ":" + "trusting"),
-        EmiStack.of(Items.SWEET_BERRIES), EMIPlugin.simplifiedRenderer(), EmiRecipeSorting.none());
+    private final TransformationRecipe recipe;
+    public static final EmiRecipeCategory TYPE = new EmiRecipeCategory(new ResourceLocation(CommonConstants.MOD_ID + ":" + "transformation"),
+            EmiStack.of(Items.GOLDEN_CARROT), EMIPlugin.simplifiedRenderer(), EmiRecipeSorting.none());
 
-    protected TrustingCategoryEMI(Builder builder, TrustingRecipe trustingRecipe) {
+    protected TransformationCategoryEMI(Builder builder, TransformationRecipe recipe) {
         super(TYPE, 168, 93, builder.id);
-        this.recipe = trustingRecipe;
+        this.recipe = recipe;
     }
 
     public static Builder builder() {
@@ -41,14 +41,14 @@ public class TrustingCategoryEMI extends AbstractRecipeCategoryEMI {
     public List<EmiIngredient> getCatalysts() {
         return List.of(EmiIngredient.of(recipe.inputStack),
                 EmiIngredient.of(recipe.extraInputStack),
-                EmiIngredient.of(recipe.spawnEgg)
+                EmiIngredient.of(recipe.inputSpawnEgg)
         );
     }
 
     @Override
     public List<EmiStack> getOutputs() {
         List<EmiStack> list = new ArrayList<>();
-        for (ItemStack item : recipe.spawnEgg.getItems()) {
+        for (ItemStack item : recipe.outputSpawnEgg.getItems()) {
             list.add(EmiStack.of(item));
         }
         return list;
@@ -56,31 +56,30 @@ public class TrustingCategoryEMI extends AbstractRecipeCategoryEMI {
 
     @Override
     public void addWidgets(WidgetHolder widgets) {
-        widgets.addSlot(EmiIngredient.of(recipe.spawnEgg), 149, 1);
+        widgets.addSlot(EmiIngredient.of(recipe.inputSpawnEgg), 65, 74);
+        widgets.addSlot(EmiIngredient.of(recipe.outputSpawnEgg), 85, 74);
 
-        int inputX = 69 + 5;
-        int inputY = 58 - 10;
-        int extraY = inputY - 19;
         boolean hasExtraInput = recipe.extraInputStack != null && !recipe.extraInputStack.isEmpty();
+        widgets.addSlot(EmiIngredient.of(recipe.inputStack), hasExtraInput ? 65 : 75, 22);
         if (hasExtraInput) {
-            widgets.addSlot(EmiIngredient.of(recipe.inputStack), inputX + 33, inputY + 9);
-            widgets.addSlot(EmiIngredient.of(recipe.extraInputStack), inputX + 33, extraY + 9);
-        } else {
-            widgets.addSlot(EmiIngredient.of(recipe.inputStack), inputX  + 33, inputY);
+            widgets.addSlot(EmiIngredient.of(recipe.extraInputStack), 85, 22);
         }
 
+        widgets.addTexture(EmiTexture.EMPTY_ARROW, 72, 49);
         EMIUtils.drawMobSlot(0, 10, widgets);
-        EMIUtils.drawMobNameAndEntity(recipe.entityType, widgets, recipe, 99, 0);
+        EMIUtils.drawMobNameAndEntity(recipe.inputEntityType, widgets, recipe, 99, 0, true, recipe.inputColor);
+        EMIUtils.drawMobSlot(105, 10, widgets);
+        EMIUtils.drawMobNameAndEntity(recipe.outputEntityType, widgets, recipe, 61, 105, false, recipe.outputColor);
     }
 
     public static class Builder {
         private ResourceLocation id;
-        private TrustingRecipe recipe;
+        private TransformationRecipe recipe;
 
         private Builder() {}
 
         public EmiRecipe build() {
-            return new TrustingCategoryEMI(this, recipe);
+            return new TransformationCategoryEMI(this, recipe);
         }
 
         public Builder id(ResourceLocation id) {
@@ -88,7 +87,7 @@ public class TrustingCategoryEMI extends AbstractRecipeCategoryEMI {
             return this;
         }
 
-        public Builder recipe(TrustingRecipe recipe) {
+        public Builder recipe(TransformationRecipe recipe) {
             this.recipe = recipe;
             return this;
         }